tests/wpt/tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-none-none.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<!-- This should behave the same as one `none` case and the case when no policy specified. -->
<meta http-equiv="Content-Security-Policy" content="trusted-types 'none' 'none'">
<body>
<script>
  test(t => {
    assert_throws_js(TypeError, _ => {
      window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } );
    });
  }, "Cannot create policy with name 'SomeName' - policy creation throws");

  test(t => {
    assert_throws_js(TypeError, _ => {
      window.trustedTypes.createPolicy('default', { createHTML: s => s } );
    });
  }, "Cannot create policy with name 'default' - policy creation throws");
</script>
